dynamic Icon based on weather

+++ lib/models/weather_model.dart
@@ -140,7 +140,7 @@ class Lista {
         weather.add(Weather.fromJson(v));
       });
     }
-    speed = json['speed'];
+    speed = json['speed'].toDouble();
     deg = json['deg'];
     gust = json['gust'].toDouble();
     clouds = json['clouds'];
